+++ modules/cache/shared/src/main/scala/free/hashmap/HashMapWrapper.scala
