+++ modules/cache/shared/src/main/scala/free/hashmap/Hasher.scala
