+++ modules/cache/shared/src/test/scala/free/CacheTests.scala
