+++ modules/integrations/fetch/shared/src/main/scala/free/fetch.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package freestyle
+package freestyle.free
 
 import _root_.fetch._
