+++ modules/integrations/slick/src/main/scala/free/slick.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package freestyle
+package freestyle.free
 
 import _root_.slick.dbio.{DBIO, DBIOAction}
 import _root_.slick.jdbc.JdbcBackend
-import freestyle.async._
+import freestyle.free.async._
 
 import scala.util.{Failure, Success}
